cluster: move topic properties out of types.h
The file is huge and means that we are required to include v_cluster if
we want to use the topic_manifest which just serializes the topic
properties.

This commit moves properties to their own file. A subsequent commit will
make it its own independent library.
